Krige soil data following the methodology of the John et al. (2007).

If called without inputs, GetKrigedSoil() calculates a "best" semivariogram to use and krige the soil data. By default the best kriging parameters (found via variogram functions in the geoR package) are used, but specified parameters can be used via the krigeParams argument. geoR has two main kriging functions: geoR::ksline() and geoR::krige.conv(). The argument useKsLine specifies whether to use the geoR::ksline() function or not.

GetKrigedSoil(df.soil, var, gridSize = 20, krigeParams = NULL,
  xSize = 1000, ySize = 500, breaks = ExpList(2, 320, 30),
  useKsLine = TRUE)

Arguments

df.soil
The data frame with the points, coords specified in the columns gx, gy.

var
A text string giving the variable/column in df.soil to krige.

gridSize
Points are kriged to the centre points of a grid of this size.

krigeParams
If you want to pass specified kriging parameters; see GetAutomatedKrigeParams() for each parameter.

xSize, ySize
X and Y size/length of the plot.

breaks
Breaks/intervals used to calculate the semivariogram, which only happens if krigeParams = NULL (default).

useKsLine
See above.

Value

A list with the following items:

df: Data frame of kriged values (column z) at each grid point (x, y).

df.poly: Data frame of the polynomial surface fitted to the raw data.

lambda: The "lambda" value used in the Box-Cox transform of the raw data.

vg: A list giving the variogram parameters used for the kriging.

vm: Minimum loss value returned from geoR::variofit().

See also

geoR::variofit(), geoR::variog(), geoR::as.geodata(), geoR::ksline(), geoR::krige.conv(), geoR::krige.control().

Examples

# NOT RUN {
# You need ggplot2
if (!requireNamespace("ggplot2", quietly = TRUE)) {
  stop("ggplot2 needed for this example to work. Please install it.",
    call. = FALSE)
}

library(ggplot2)

# Use randomized data set for examples (invalid for reaserch).
df <- fgeo.habitat::soil_random
str(df)

# Krige with automated parameters
kriged <- GetKrigedSoil(df, var = "M3Al")
str(kriged)

# Krige with specified parameters - these example params are rather arbitrary
# but are loosely based on what was chosen from the automated kriging
params <- list(
  model = "circular",
  range = 100,
  nugget = 1000,
  sill = 46000,
  kappa = 0.5
)
kriged.2 <- GetKrigedSoil(df, var = "M3Al", krigeParams = params)

# Have a look at the differences
g <- ggplot(kriged$df, aes(x = x, y = y, fill = z)) +
  geom_tile() +
  coord_equal()
g2 <- ggplot(kriged.2$df, aes(x = x, y = y, fill = z)) +
  geom_tile() +
  coord_equal()
g
g2
# }