Krige soil data following the methodology of the John et al. (2007).

Functions to krige soil data:

GetKrigedSoil() is the original function but it is softly deprecated. Its result is the same as krig()[[1]] when run for one soil variable.

krig() some additional convenient features:

An interface consistent with other functions of ForestGEO.

A a method for summary()(see examples).

Tries to guess plotdim.

Informs the guessed plotdim and the gridsize provided.

Allows to suppress messages.

Iterating over multiple soil variables.

Has a method for to_df() (see examples).

[[1]: R:[1

krig(soil, var, params = NULL, gridsize = 20,
  plotdim = guess_plotdim(soil), breaks = krig_breaks(2, 320, 30),
  use_ksline = TRUE, quiet = FALSE)

GetKrigedSoil(df.soil, var, gridSize = 20, krigeParams = NULL,
  xSize = 1000, ySize = 500, breaks = krig_breaks(2, 320, 30),
  useKsLine = TRUE)

Arguments

soil, df.soil
The data frame with the points, coords specified in the columns gx, gy.

var
A character vector giving the name in the soil dataset of the variable(s) containing soil data to krige. var must be of length 1 for GetKrigedSoil() but can be of length greater than one for krig().

params, krigeParams
If you want to pass specified kriging parameters; see krig_auto_params() for each parameter.

gridsize, gridSize
Points are kriged to the center points of a grid of this size.

plotdim, xSize, ySize
Numeric vectors giving x and y dimensions of the plot:

plotdim: Must be of length 2 with the format c(x, y).

xSize, ySize: Each must be of length 1.

breaks
Breaks/intervals used to calculate the semivariogram, which only happens if krigeParams = NULL (default).

use_ksline, useKsLine
Use the geoR::ksline() function? Use TRUE to calculate a "best" semivariogram based on default parameters via geoR::variogram()]. Use FALSE to base calculation on parameters passed to params.

quiet
Use TRUE to suppresses messages.

Value

A list with the following items:

df: Data frame of kriged values (column z) at each grid point (x, y).

df.poly: Data frame of the polynomial surface fitted to the raw data.

lambda: The "lambda" value used in the Box-Cox transform of the raw data.

vg: A list giving the variogram parameters used for the kriging.

vm: Minimum loss value returned from geoR::variofit().

Breaks default

The default breaks argument is set to have more points where the exponential curve rises the most and fewer at large distances. This means that the curve fitting is not overly biased by points beyond the effective maximum range.

See also

Examples

# Original funciton
# Using automated parameters
auto <- GetKrigedSoil(soil_fake, var = "c")

#> variog: computing omnidirectional variogram
#> variofit: covariance model used is exponential 
#> variofit: weights used: npairs 
#> variofit: minimisation function used: optim 
#> variofit: covariance model used is circular 
#> variofit: weights used: npairs 
#> variofit: minimisation function used: optim 
#> variofit: covariance model used is cauchy 
#> variofit: weights used: npairs 
#> variofit: minimisation function used: optim 
#> variofit: covariance model used is gaussian 
#> variofit: weights used: npairs 
#> variofit: minimisation function used: optim 
#> ksline: kriging location:  1 out of 1250 
#> ksline: kriging location:  101 out of 1250 
#> ksline: kriging location:  201 out of 1250 
#> ksline: kriging location:  301 out of 1250 
#> ksline: kriging location:  401 out of 1250 
#> ksline: kriging location:  501 out of 1250 
#> ksline: kriging location:  601 out of 1250 
#> ksline: kriging location:  701 out of 1250 
#> ksline: kriging location:  801 out of 1250 
#> ksline: kriging location:  901 out of 1250 
#> ksline: kriging location:  1001 out of 1250 
#> ksline: kriging location:  1101 out of 1250 
#> ksline: kriging location:  1201 out of 1250 
#> ksline: kriging location:  1250 out of 1250 
#> Kriging performed using global neighbourhood 

summary(auto)

#>         Length Class      Mode   
#> df       3     data.frame list   
#> df.poly  3     data.frame list   
#> lambda   1     -none-     numeric
#> vg      20     variogram  list   
#> vm      17     variomodel list   

# Wrapper produces the same output and add convenient features

# Now using custom parameters (arbitrary but based on automated kriging params)
params <- list(
  model = "circular", range = 100, nugget = 1000, sill = 46000, kappa = 0.5
)

# Also using not one but multiple soil variables
vars <- c("c", "p")
custom <- krig(soil_fake, vars, params = params, quiet = TRUE)
summary(custom)

#> var: c 
#> df
#> 'data.frame':	1150 obs. of  3 variables:
#>  $ x: num  10 30 50 70 90 110 130 150 170 190 ...
#>  $ y: num  10 10 10 10 10 10 10 10 10 10 ...
#>  $ z: num  2.29 2.31 2.22 2.04 1.79 ...
#> 
#> df.poly
#> 'data.frame':	1150 obs. of  3 variables:
#>  $ gx: num  10 30 50 70 90 110 130 150 170 190 ...
#>  $ gy: num  10 10 10 10 10 10 10 10 10 10 ...
#>  $ z : num  2.13 2.12 2.1 2.09 2.07 ...
#> 
#> lambda
#> 'numeric'
#>  num 1
#> 
#> vg
#> 'NULL'
#>  NULL
#> 
#> vm
#> 'NULL'
#>  NULL
#> 
#> var: p 
#> df
#> 'data.frame':	1150 obs. of  3 variables:
#>  $ x: num  10 30 50 70 90 110 130 150 170 190 ...
#>  $ y: num  10 10 10 10 10 10 10 10 10 10 ...
#>  $ z: num  6.2 6.09 6.01 6.03 6.1 ...
#> 
#> df.poly
#> 'data.frame':	1150 obs. of  3 variables:
#>  $ gx: num  10 30 50 70 90 110 130 150 170 190 ...
#>  $ gy: num  10 10 10 10 10 10 10 10 10 10 ...
#>  $ z : num  6.37 6.35 6.33 6.31 6.29 ...
#> 
#> lambda
#> 'numeric'
#>  num 1
#> 
#> vg
#> 'NULL'
#>  NULL
#> 
#> vm
#> 'NULL'
#>  NULL
#> 

result <- to_df(custom, item = "df")
head(result)

#>   var   x  y        z
#> 1   c  10 10 2.290586
#> 2   c  30 10 2.306093
#> 3   c  50 10 2.221311
#> 4   c  70 10 2.037240
#> 5   c  90 10 1.793122
#> 6   c 110 10 1.544911

tail(result)

#>      var   x   y        z
#> 2295   p 890 450 5.843585
#> 2296   p 910 450 5.835235
#> 2297   p 930 450 5.827757
#> 2298   p 950 450 5.821150
#> 2299   p 970 450 5.815414
#> 2300   p 990 450 5.810549