Convert a matrices to data frames.
Source: R/mat_enframe.R

These functions convert a matrix to three-column data frames, where columns store row names, column names, and values of the original matrix. mat_enframe() outputs a long-format dataframe and mat_enframe_ls() returns a list of dataframes where each element of the list maps to a column of the original matrix.

mat_enframe(...)

mat_enframe_ls(mat, rownm = "rownames", colnm = "colnames",
  value = "value")

Arguments

...
Arguments passed to mat_enframe_ls.

mat
A matrix.

rownm, colnm, value
Names of the columns that store the row names, the column names and the values.

Value

mat_enframe(): A dataframe.

mat_enframe_ls(): A list of dataframes.

Details

These functions are a matrix implementation of the enframe() function of the tibble package.

See also

enframe() (tibble package).

Examples

mat <- matrix(1:6, 2, dimnames = list(LETTERS[1:2], letters[1:3]))
mat_enframe_ls(mat)

#> $a
#>   rownames colnames value
#> 1        A        a     1
#> 2        B        a     2
#> 
#> $b
#>   rownames colnames value
#> 1        A        b     3
#> 2        B        b     4
#> 
#> $c
#>   rownames colnames value
#> 1        A        c     5
#> 2        B        c     6
#> 

mat_enframe_ls(mat, "metric", "sp")

#> $a
#>   metric sp value
#> 1      A  a     1
#> 2      B  a     2
#> 
#> $b
#>   metric sp value
#> 1      A  b     3
#> 2      B  b     4
#> 
#> $c
#>   metric sp value
#> 1      A  c     5
#> 2      B  c     6
#> 

mat <- matrix(1:6, 2)
mat_enframe_ls(mat)

#> $X1
#>   rownames colnames value
#> 1        1        1     1
#> 2        2        1     2
#> 
#> $X2
#>   rownames colnames value
#> 1        1        2     3
#> 2        2        2     4
#> 
#> $X3
#>   rownames colnames value
#> 1        1        3     5
#> 2        2        3     6
#>