Torus Translation Test to determine habitat associations of tree species.

Torus Translation Test (TT test) to determine habitat associations of tree species. Based on code written by Kyle Harms.

torusonesp.all(species, hab.index20, allabund20, plotdim, gridsize)

Arguments

species
Character sting giving the name of one species.

hab.index20
Object giving the habitat designation for each 20x20 quadrat.

allabund20
Abundance per quadrat.

plotdim
Plot dimensions.

gridsize
Grid size. Make sure the gridsize entered here matches the gridsize on which the habitats are defined and the abundances were calculated

Value

A numeric matrix.

Details

You should only try to determine the habitat association for sufficiently abundant species - in a 50-ha plot, a minimum abundance of 50 trees/species has been used.

See also

Example at https://bookdown.org/fgeocomm/ttt/.

Examples

# Small dataset from Luquillo
census_data <- luquillo_top3_sp
alive_trees <- census_data[census_data$status == "A", ]

habitat_data <- luquillo_habitat
plot_dimensions <- c(320, 500)
grid_size <- 20

abundance_per_quadrat <- abundanceperquad(
  alive_trees,
  plotdim = plot_dimensions,
  gridsize = grid_size,
  type = 'abund'
)$abund

one_species <- unique(census_data$sp)[[1]]
out_one <- torusonesp.all(
  species = one_species,
  hab.index20 = habitat_data,
  allabund20 = abundance_per_quadrat,
  plotdim = plot_dimensions,
  gridsize = grid_size
)
t(out_one)

#>                     CASARB
#> N.Hab.1          25.000000
#> Gr.Hab.1       1489.000000
#> Ls.Hab.1        109.000000
#> Eq.Hab.1          2.000000
#> Rep.Agg.Neut.1    0.000000
#> Obs.Quantile.1    0.930625
#> N.Hab.2          12.000000
#> Gr.Hab.2        168.000000
#> Ls.Hab.2       1431.000000
#> Eq.Hab.2          1.000000
#> Rep.Agg.Neut.2    0.000000
#> Obs.Quantile.2    0.105000
#> N.Hab.3          14.000000
#> Gr.Hab.3        567.000000
#> Ls.Hab.3       1029.000000
#> Eq.Hab.3          4.000000
#> Rep.Agg.Neut.3    0.000000
#> Obs.Quantile.3    0.354375
#> N.Hab.4          15.000000
#> Gr.Hab.4        934.000000
#> Ls.Hab.4        661.000000
#> Eq.Hab.4          5.000000
#> Rep.Agg.Neut.4    0.000000
#> Obs.Quantile.4    0.583750

# Iterate over multiple species
all_species <- unique(census_data$sp)
out_all <- lapply(
  X = all_species,
  FUN = torusonesp.all,
  # Other arguments passed to torusonesp.all
  hab.index20 = habitat_data,
  allabund20 = abundance_per_quadrat,
  plotdim = plot_dimensions,
  gridsize = grid_size
)
# Compact view
t(Reduce(rbind, out_all))

#>                     CASARB      PREMON      SLOBER
#> N.Hab.1          25.000000   59.000000   14.000000
#> Gr.Hab.1       1489.000000  389.000000  492.000000
#> Ls.Hab.1        109.000000 1208.000000 1092.000000
#> Eq.Hab.1          2.000000    3.000000   16.000000
#> Rep.Agg.Neut.1    0.000000    0.000000    0.000000
#> Obs.Quantile.1    0.930625    0.243125    0.307500
#> N.Hab.2          12.000000   75.000000   16.000000
#> Gr.Hab.2        168.000000 1562.000000  473.000000
#> Ls.Hab.2       1431.000000   37.000000 1125.000000
#> Eq.Hab.2          1.000000    1.000000    2.000000
#> Rep.Agg.Neut.2    0.000000    1.000000    0.000000
#> Obs.Quantile.2    0.105000    0.976250    0.295625
#> N.Hab.3          14.000000   56.000000   19.000000
#> Gr.Hab.3        567.000000  632.000000 1181.000000
#> Ls.Hab.3       1029.000000  963.000000  415.000000
#> Eq.Hab.3          4.000000    5.000000    4.000000
#> Rep.Agg.Neut.3    0.000000    0.000000    0.000000
#> Obs.Quantile.3    0.354375    0.395000    0.738125
#> N.Hab.4          15.000000   44.000000   17.000000
#> Gr.Hab.4        934.000000  222.000000 1151.000000
#> Ls.Hab.4        661.000000 1375.000000  440.000000
#> Eq.Hab.4          5.000000    3.000000    9.000000
#> Rep.Agg.Neut.4    0.000000    0.000000    0.000000
#> Obs.Quantile.4    0.583750    0.138750    0.719375