Torus Translation Test to determine habitat associations of tree species.

Torus Translation Test (TT test) to determine habitat associations of tree species.

tt_test(sp, habitat, abundance, plotdim = extract_plotdim(habitat),
  gridsize = extract_gridsize(habitat))

torusonesp.all(species, hab.index20, allabund20, plotdim, gridsize)

Arguments

sp, species
Character sting giving species names. tt_test() and torusonesp.all()can take only one species; tt_test_lst() can take any number of species.

habitat, hab.index20
Object giving the habitat designation for each plot partition defined by gridsize.

abundance, allabund20
The output of abund_index().

plotdim
Plot dimensions.

gridsize
Grid size. If using tt_test(), ensure it matches the gridsize on which the habitats are defined and the abundances were calculated.

census
A dataframe; a ForestGEO census.

Value

tt_test_lst(): A dataframe.

tt_test(): A numeric matrix.

Details

You should only try to determine the habitat association for sufficiently abundant species - in a 50-ha plot, a minimum abundance of 50 trees/species has been used.

tt_test_lst() uses abundanceperquad() -- via abund_index() -- which is slow. You may calculate abundance per quadrat independently, feed it to the argument abundance of tt_test(), and reformat the output with as_df(). You can iterate over multiple species with a for loop or a functional such as lapply.

See also

Examples

# For easier data wranging
library(dplyr)

habitat <- luquillo_habitat
census <- luquillo_top3_sp

# Pick alive trees, of 10 mm or more
pick <- filter(census, status == "A", dbh >= 10)
# Pick sufficiently abundant trees
pick <- add_count(pick, sp)
pick <- filter(pick, n > 50)

species <- unique(pick$sp)

# Test any number of species (output a list of matrices)
tt_lst <- tt_test_lst(census, species, habitat)
str(tt_lst, give.attr = FALSE)

#> List of 3
#>  $ : 'tt' num [1, 1:24] 25 1489 109 2 0 ...
#>  $ : 'tt' num [1, 1:24] 59 389 1208 3 0 ...
#>  $ : 'tt' num [1, 1:24] 14 492 1092 16 0 ...

tt_lst[[1]]

#>        N.Hab.1 Gr.Hab.1 Ls.Hab.1 Eq.Hab.1 Rep.Agg.Neut.1 Obs.Quantile.1 N.Hab.2
#> CASARB      25     1489      109        2              0          0.931      12
#>        Gr.Hab.2 Ls.Hab.2 Eq.Hab.2 Rep.Agg.Neut.2 Obs.Quantile.2 N.Hab.3
#> CASARB      168     1431        1              0          0.105      14
#>        Gr.Hab.3 Ls.Hab.3 Eq.Hab.3 Rep.Agg.Neut.3 Obs.Quantile.3 N.Hab.4
#> CASARB      567     1029        4              0          0.354      15
#>        Gr.Hab.4 Ls.Hab.4 Eq.Hab.4 Rep.Agg.Neut.4 Obs.Quantile.4
#> CASARB      934      661        5              0          0.584
#> attr(,"class")
#> [1] "tt"     "matrix"

# Try also: View((as_df(tt_lst)))
head(as_df(tt_lst))

#>           metric     sp    value
#> 1        N.Hab.1 CASARB   25.000
#> 2       Gr.Hab.1 CASARB 1489.000
#> 3       Ls.Hab.1 CASARB  109.000
#> 4       Eq.Hab.1 CASARB    2.000
#> 5 Rep.Agg.Neut.1 CASARB    0.000
#> 6 Obs.Quantile.1 CASARB    0.931

# Iterate over multiple species
plotdim <- c(320, 500)
gridsize <- 20
abundance <- abund_index(pick, plotdim, gridsize)

tt_lst <- lapply(species, tt_test, habitat, abundance, plotdim, gridsize)
tt_mat

#>        N.Hab.1 Gr.Hab.1 Ls.Hab.1 Eq.Hab.1 Rep.Agg.Neut.1 Obs.Quantile.1 N.Hab.2
#> PREMON      59      389     1208        3              0          0.243      75
#>        Gr.Hab.2 Ls.Hab.2 Eq.Hab.2 Rep.Agg.Neut.2 Obs.Quantile.2 N.Hab.3
#> PREMON     1562       37        1              1          0.976      56
#>        Gr.Hab.3 Ls.Hab.3 Eq.Hab.3 Rep.Agg.Neut.3 Obs.Quantile.3 N.Hab.4
#> PREMON      632      963        5              0          0.395      44
#>        Gr.Hab.4 Ls.Hab.4 Eq.Hab.4 Rep.Agg.Neut.4 Obs.Quantile.4
#> PREMON      222     1375        3              0          0.139

# Test one species with original function (outputs a matrix)
tt_mat <- torusonesp.all(species[[1]],
  hab.index20 = habitat,
  allabund20 = abundance,
  plotdim = plotdim,
  gridsize = gridsize
)
tt_mat

#>        N.Hab.1 Gr.Hab.1 Ls.Hab.1 Eq.Hab.1 Rep.Agg.Neut.1 Obs.Quantile.1 N.Hab.2
#> CASARB      25     1489      109        2              0          0.931      12
#>        Gr.Hab.2 Ls.Hab.2 Eq.Hab.2 Rep.Agg.Neut.2 Obs.Quantile.2 N.Hab.3
#> CASARB      168     1431        1              0          0.105      14
#>        Gr.Hab.3 Ls.Hab.3 Eq.Hab.3 Rep.Agg.Neut.3 Obs.Quantile.3 N.Hab.4
#> CASARB      567     1029        4              0          0.354      15
#>        Gr.Hab.4 Ls.Hab.4 Eq.Hab.4 Rep.Agg.Neut.4 Obs.Quantile.4
#> CASARB      934      661        5              0          0.584