Krige soil data following the methodology of the John et al. (2007).

Functions to krige soil data. The most common usage patterns are summary(krig(ONE-SOIL-VARIABLE, ...)) and to_df(krig_lst(ANY-NUMBER-OF-SOIL-VARIABLES, ...)); but you may use other options:

GetKrigedSoil() is the original function but it is softly deprecated. Its result is the same as krig() but its interface is different and lacks some convenient features.

krig() has an interface consistent with other functions of ForestGEO. It outputs a list of subclass "krig", which has a method for summary()(see examples). krig() also tries to guess plotdim and informs both the guessed plotdim and the gridsize provided. It also allows you to suppress messages.

krig_lst allows iterating over multiple soil variables. It lists the result of each iteration and adds the subclass "krig_lst" which has a method for to_df() (see examples).

krig(soil, var, params = NULL, gridsize = 20,
  plotdim = guess_plotdim(soil), breaks = krig_breaks(2, 320, 30),
  use_ksline = TRUE, quiet = FALSE)

krig_lst(soil, var, params = NULL, gridsize = 20,
  plotdim = guess_plotdim(soil), breaks = krig_breaks(2, 320, 30),
  use_ksline = TRUE, quiet = FALSE)

GetKrigedSoil(df.soil, var, gridSize = 20, krigeParams = NULL,
  xSize = 1000, ySize = 500, breaks = krig_breaks(2, 320, 30),
  useKsLine = TRUE)

Arguments

soil, df.soil
The data frame with the points, coords specified in the columns gx, gy.

var
A character vector giving the name in the soil dataset of the variable(s) containing soil data to krige. var must be of length 1 for krig() but can be of length greater than one for krig_lst().

params, krigeParams
If you want to pass specified kriging parameters; see krig_auto_params() for each parameter.

gridsize, gridSize
Points are kriged to the center points of a grid of this size.

plotdim, xSize, ySize
Numeric vectors giving x and y dimensions of the plot:

plotdim: Must be of length 2 with the format c(x, y).

xSize, ySize: Each must be of length 1.

breaks
Breaks/intervals used to calculate the semivariogram, which only happens if krigeParams = NULL (default).

use_ksline, useKsLine
Use the geoR::ksline() function? Use TRUE to calculate a "best" semivariogram based on default parameters via geoR::variogram()]. Use FALSE to base calculation on parameters passed to params.

quiet
Use TRUE to suppresses messages.

...
Arguments passed to krig().

Value

A list with the following items:

df: Data frame of kriged values (column z) at each grid point (x, y).

df.poly: Data frame of the polynomial surface fitted to the raw data.

lambda: The "lambda" value used in the Box-Cox transform of the raw data.

vg: A list giving the variogram parameters used for the kriging.

vm: Minimum loss value returned from geoR::variofit().

Breaks default

The default breaks argument is set to have more points where the exponential curve rises the most and fewer at large distances. This means that the curve fitting is not overly biased by points beyond the effective maximum range.

See also

Examples

# Example data
soil <- soil_random

# Krige with automated parameters

# Original funciton
auto1 <- GetKrigedSoil(soil, var = "m3al")

#> variog: computing omnidirectional variogram
#> variofit: covariance model used is exponential 
#> variofit: weights used: npairs 
#> variofit: minimisation function used: optim 
#> variofit: covariance model used is circular 
#> variofit: weights used: npairs 
#> variofit: minimisation function used: optim 
#> variofit: covariance model used is cauchy 
#> variofit: weights used: npairs 
#> variofit: minimisation function used: optim 
#> variofit: covariance model used is gaussian 
#> variofit: weights used: npairs 
#> variofit: minimisation function used: optim 
#> ksline: kriging location:  1 out of 1250 
#> ksline: kriging location:  101 out of 1250 
#> ksline: kriging location:  201 out of 1250 
#> ksline: kriging location:  301 out of 1250 
#> ksline: kriging location:  401 out of 1250 
#> ksline: kriging location:  501 out of 1250 
#> ksline: kriging location:  601 out of 1250 
#> ksline: kriging location:  701 out of 1250 
#> ksline: kriging location:  801 out of 1250 
#> ksline: kriging location:  901 out of 1250 
#> ksline: kriging location:  1001 out of 1250 
#> ksline: kriging location:  1101 out of 1250 
#> ksline: kriging location:  1201 out of 1250 
#> ksline: kriging location:  1250 out of 1250 
#> Kriging performed using global neighbourhood 

summary(auto1)

#>         Length Class      Mode   
#> df       3     data.frame list   
#> df.poly  3     data.frame list   
#> lambda   1     -none-     numeric
#> vg      20     variogram  list   
#> vm      17     variomodel list   

# Wrappers produce the same output but have some convenient feature
# Suppress messages
auto2 <- krig(soil, var = "m3al", quiet = TRUE)
# More complete summary
summary(auto1)

#>         Length Class      Mode   
#> df       3     data.frame list   
#> df.poly  3     data.frame list   
#> lambda   1     -none-     numeric
#> vg      20     variogram  list   
#> vm      17     variomodel list   

# Enable more compleate summary with output of GetKrigedSoil()
summary(as_krig(auto1))

#> df
#> 'data.frame':	1250 obs. of  3 variables:
#>  $ x: num  10 30 50 70 90 110 130 150 170 190 ...
#>  $ y: num  10 10 10 10 10 10 10 10 10 10 ...
#>  $ z: num  950 944 939 933 928 ...
#> 
#> df.poly
#> 'data.frame':	1250 obs. of  3 variables:
#>  $ gx: num  10 30 50 70 90 110 130 150 170 190 ...
#>  $ gy: num  10 10 10 10 10 10 10 10 10 10 ...
#>  $ z : num  949 943 938 933 927 ...
#> 
#> lambda
#> 'numeric'
#>  num 1
#> 
#> vg
#> 'variogram'
#> List of 20
#>  $ u               : num [1:12] 30.3 42.9 51.1 60.9 86.5 ...
#>  $ v               : num [1:12] 60522 54239 174166 55226 44193 ...
#>  $ n               : num [1:12] 21 25 5 81 94 57 155 167 247 288 ...
#>  $ sd              : num [1:12] 96683 58618 214627 60326 52526 ...
#>  $ bins.lim        : num [1:31] 1.00e-12 2.00 2.38 2.84 3.38 ...
#>  $ ind.bin         : logi [1:30] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  $ var.mark        : num 55039
#>  $ beta.ols        : num 2.62e-08
#>  $ output.type     : chr "bin"
#>  $ max.dist        : num 320
#>  $ estimator.type  : chr "classical"
#>  $ n.data          : int 100
#>  $ lambda          : num 1
#>  $ trend           : chr "cte"
#>  $ pairs.min       : num 5
#>  $ nugget.tolerance: num 1e-12
#>  $ direction       : chr "omnidirectional"
#>  $ tolerance       : chr "none"
#>  $ uvec            : num [1:30] 1 2.19 2.61 3.11 3.7 ...
#>  $ call            : language variog(geodata = geodata, breaks = breaks, trend = trend, pairs.min = 5)
#> 
#> vm
#> 'variomodel', variofit'
#> List of 17
#>  $ nugget               : num 56634
#>  $ cov.pars             : num [1:2] 87783 2105
#>  $ cov.model            : chr "gaussian"
#>  $ kappa                : num 0.5
#>  $ value                : num 1.49e+11
#>  $ trend                : chr "cte"
#>  $ beta.ols             : num 2.62e-08
#>  $ practicalRange       : num 3644
#>  $ max.dist             : num 320
#>  $ minimisation.function: chr "optim"
#>  $ weights              : chr "npairs"
#>  $ method               : chr "WLS"
#>  $ fix.nugget           : logi FALSE
#>  $ fix.kappa            : logi TRUE
#>  $ lambda               : num 1
#>  $ message              : chr "optim convergence code: 0"
#>  $ call                 : language variofit(vario = vg, ini.cov.pars = c(initialVal, startRange), cov.model = varModels[i],      nugget = initialVal)

# Krige with custom params (arbitrary but based on automated kriging params)

params <- list(
  model = "circular", range = 100, nugget = 1000, sill = 46000, kappa = 0.5
)
custom <- krig(soil, var = "m3al", params = params, quiet = TRUE)
summary(custom)

#> df
#> 'data.frame':	1250 obs. of  3 variables:
#>  $ x: num  10 30 50 70 90 110 130 150 170 190 ...
#>  $ y: num  10 10 10 10 10 10 10 10 10 10 ...
#>  $ z: num  933 955 984 968 919 ...
#> 
#> df.poly
#> 'data.frame':	1250 obs. of  3 variables:
#>  $ gx: num  10 30 50 70 90 110 130 150 170 190 ...
#>  $ gy: num  10 10 10 10 10 10 10 10 10 10 ...
#>  $ z : num  949 943 938 933 927 ...
#> 
#> lambda
#> 'numeric'
#>  num 1
#> 
#> vg
#> 'NULL'
#>  NULL
#> 
#> vm
#> 'NULL'
#>  NULL

# Krig multiple soil variables

soil <- soil_fake
vars <- c("c", "p")
out_lst <- krig_lst(soil, vars, quiet = TRUE)

result <- to_df(out_lst, item = "df")
head(result)

#>   var   x  y     z
#> 1   c  10 10 0.837
#> 2   c  30 10 0.811
#> 3   c  50 10 0.785
#> 4   c  70 10 0.759
#> 5   c  90 10 0.733
#> 6   c 110 10 0.707

tail(result)

#>      var   x   y      z
#> 2295   p 890 450 -0.598
#> 2296   p 910 450 -0.613
#> 2297   p 930 450 -0.626
#> 2298   p 950 450 -0.639
#> 2299   p 970 450 -0.649
#> 2300   p 990 450 -0.658

# Same with original function

result2 <- lapply(vars, function(.vars) GetKrigedSoil(soil, .vars))

#> variog: computing omnidirectional variogram
#> variofit: covariance model used is exponential 
#> variofit: weights used: npairs 
#> variofit: minimisation function used: optim 
#> variofit: covariance model used is circular 
#> variofit: weights used: npairs 
#> variofit: minimisation function used: optim 
#> variofit: covariance model used is cauchy 
#> variofit: weights used: npairs 
#> variofit: minimisation function used: optim 
#> variofit: covariance model used is gaussian 
#> variofit: weights used: npairs 
#> variofit: minimisation function used: optim 
#> ksline: kriging location:  1 out of 1250 
#> ksline: kriging location:  101 out of 1250 
#> ksline: kriging location:  201 out of 1250 
#> ksline: kriging location:  301 out of 1250 
#> ksline: kriging location:  401 out of 1250 
#> ksline: kriging location:  501 out of 1250 
#> ksline: kriging location:  601 out of 1250 
#> ksline: kriging location:  701 out of 1250 
#> ksline: kriging location:  801 out of 1250 
#> ksline: kriging location:  901 out of 1250 
#> ksline: kriging location:  1001 out of 1250 
#> ksline: kriging location:  1101 out of 1250 
#> ksline: kriging location:  1201 out of 1250 
#> ksline: kriging location:  1250 out of 1250 
#> Kriging performed using global neighbourhood 
#> variog: computing omnidirectional variogram
#> variofit: covariance model used is exponential 
#> variofit: weights used: npairs 
#> variofit: minimisation function used: optim 
#> variofit: covariance model used is circular 
#> variofit: weights used: npairs 
#> variofit: minimisation function used: optim 
#> variofit: covariance model used is cauchy 
#> variofit: weights used: npairs 
#> variofit: minimisation function used: optim 
#> variofit: covariance model used is gaussian 
#> variofit: weights used: npairs 
#> variofit: minimisation function used: optim 
#> ksline: kriging location:  1 out of 1250 
#> ksline: kriging location:  101 out of 1250 
#> ksline: kriging location:  201 out of 1250 
#> ksline: kriging location:  301 out of 1250 
#> ksline: kriging location:  401 out of 1250 
#> ksline: kriging location:  501 out of 1250 
#> ksline: kriging location:  601 out of 1250 
#> ksline: kriging location:  701 out of 1250 
#> ksline: kriging location:  801 out of 1250 
#> ksline: kriging location:  901 out of 1250 
#> ksline: kriging location:  1001 out of 1250 
#> ksline: kriging location:  1101 out of 1250 
#> ksline: kriging location:  1201 out of 1250 
#> ksline: kriging location:  1250 out of 1250 
#> Kriging performed using global neighbourhood 

# To enable convenient features first coerce the result to a krig_lst object
result2 <- setNames(result2, vars)
coerced <- as_krig_lst(result2)
head(to_df(coerced))

#>   var   x  y     z
#> 1   c  10 10 0.837
#> 2   c  30 10 0.811
#> 3   c  50 10 0.785
#> 4   c  70 10 0.759
#> 5   c  90 10 0.733
#> 6   c 110 10 0.707