Krige soil data following the methodology of the John et al. (2007).

Krige soil data following the methodology of the John et al. (2007):

krig() has an interface consistent with other functions of ForestGEO. It outputs a list of subclass "krig", which has a summary() method (see examples). krig() also tries to guess plotdim and informs both the guessed plotdim and the gridsize provided. krig() allows you to suppress messages.

GetKrigedSoil() remains to preserve the original interface but it is softly deprecated. Its result is the same as krig() but its interface is different and lacks many of krig()'s features.

krig(soil, var, params = NULL, gridsize = 20,
  plotdim = guess_plotdim(soil), breaks = krig_breaks(2, 320, 30),
  use_ksline = TRUE, quiet = FALSE)

GetKrigedSoil(df.soil, var, gridSize = 20, krigeParams = NULL,
  xSize = 1000, ySize = 500, breaks = krig_breaks(2, 320, 30),
  useKsLine = TRUE)

Arguments

soil, df.soil
The data frame with the points, coords specified in the columns gx, gy.

var
A text string giving the variable/column in df.soil to krige.

params, krigeParams
If you want to pass specified kriging parameters; see krig_auto_params() for each parameter.

gridsize, gridSize
Points are kriged to the center points of a grid of this size.

plotdim, xSize, ySize
Numeric vectors giving x and y dimensions of the plot:

plotdim: Must be of lenght 2 with the format c(x, y).

xSize, ySize: Each must be of lenght 1.

breaks
Breaks/intervals used to calculate the semivariogram, which only happens if krigeParams = NULL (default).

use_ksline, useKsLine
Use the geoR::ksline() function? Use TRUE to calculate a "best" semivariogram based on default parameters via geoR::variogram()]. Use FALSE to base calculation on parameters passed to params.

Value

A list with the following items:

df: Data frame of kriged values (column z) at each grid point (x, y).

df.poly: Data frame of the polynomial surface fitted to the raw data.

lambda: The "lambda" value used in the Box-Cox transform of the raw data.

vg: A list giving the variogram parameters used for the kriging.

vm: Minimum loss value returned from geoR::variofit().

Breaks default

The default breaks argument is set to have more points where the exponential curve rises the most and fewer at large distances. This means that the curve fitting is not overly biased by points beyond the effective maximum range.

See also

Examples

# NOT RUN {
# Example data
soil <- soil_random

# Krige with automated parameters
auto <- krig(soil, var = "m3al")
# For a nice view in RStudio use: View(auto)
summary(auto)

# Custom params (arbitrary but based on automated kriging params)
params <- list(
  model = "circular", range = 100, nugget = 1000, sill = 46000, kappa = 0.5
)
custom <- krig(soil, var = "m3al", params = params)
# For a nice view in RStudio use: View(custom)
summary(custom)

# Compare
if (!requireNamespace("ggplot2", quietly = TRUE)) {
  stop("Install ggplot2 for this section to work.", call. = FALSE)
}
library(ggplot2)

title_auto <- "Automated parameters"
ggplot(auto$df, aes(x = x, y = y, fill = z)) +
  geom_tile() +
  coord_equal() +
  labs(title = title_auto)
title_custom <- "Custom parameters"

ggplot(custom$df, aes(x = x, y = y, fill = z)) +
  geom_tile() +
  coord_equal() +
  labs(title = title_custom)
# }